+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 0.1.0 (2023/10/01)
 
 - Added basic deck editing
+- Added state saving
 - Added settings
   - Hash saving
   - Config page